Fixing sigset_t type and alarm
alarm is __alarm, it was failing miserably since I did the trick
  for avoiding collision between Harvey's syscall and ANSI function.

Moving sleep.c to _sleep.c, now all underscored functions for avoiding
  collisions between native and ANSI functions are compliance with this
  style.

Fix typo in sigsuspend. Sigsuspend() now fails in notecont() like kill().
Investing signal handler.

+++ include/signal.h
@@ -63,7 +63,7 @@ extern int raise(int);
 
 #ifdef _POSIX_SOURCE
 
-typedef long sigset_t;
+typedef unsigned long long sigset_t;
 struct sigaction {
 	void		(*sa_handler)();
 	sigset_t	sa_mask;
